test pull request on test_branch
testing trivial update to sketch.js file via newly created test branch.

curious to see if
a) Daniel is notified (without my directly asking on Slack).
b) if he would be so kind to merge  this in to confirm this most basic workflow.
c) or if creating separate branch followed by PR and merging to master if overkill. (seems like a safe / good idea in this context).